refactor(filters): split filter-bar into focused subcomponents
Extract the three internal subcomponents of FilterBar into their own
files so the orchestrator stays small and each piece is independently
readable / testable:

- filter-chip.tsx — single active filter chip + the formatFilterSummary
  helper
- add-filter-popover.tsx — two-step field picker + filter editor popover
- presets-menu.tsx — dropdown of one-click filter bundles

FilterBar drops from 431 → 187 lines, and its top-level imports shrink
to the modules it actually uses (router hooks, DebouncedInput, Button,
and the three new pieces). No behavior change.
