+++ components/filters/filter-bar.tsx
@@ -1,41 +1,18 @@
 'use client'
 
-import { PlusIcon, SearchIcon, SparklesIcon, XIcon } from 'lucide-react'
+import { SearchIcon, XIcon } from 'lucide-react'
 
 import type { FilterDraft } from '@/components/filters/filter-editor'
-import type {
-  ActiveFilter,
-  FilterField,
-  FilterPreset,
-  FilterSchema,
-} from '@/lib/filters/types'
+import type { FilterPreset } from '@/lib/filters/types'
 import type { QueryConfig } from '@/types/query-config'
 
 import { usePathname, useRouter, useSearchParams } from 'next/navigation'
-import { useCallback, useMemo, useState } from 'react'
-import { FilterEditor } from '@/components/filters/filter-editor'
+import { useCallback, useMemo } from 'react'
+import { AddFilterPopover } from '@/components/filters/add-filter-popover'
+import { FilterChip } from '@/components/filters/filter-chip'
+import { PresetsMenu } from '@/components/filters/presets-menu'
 import { Button } from '@/components/ui/button'
-import {
-  Command,
-  CommandEmpty,
-  CommandGroup,
-  CommandInput,
-  CommandItem,
-  CommandList,
-} from '@/components/ui/command'
 import { DebouncedInput } from '@/components/ui/debounced-input'
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu'
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@/components/ui/popover'
-import { OPERATORS } from '@/lib/filters/operators'
 import {
   parseFiltersFromParams,
   serializeFilter,
@@ -206,226 +183,5 @@ export function FilterBar({ queryConfig }: FilterBarProps) {
         </Button>
       )}
     </div>
-  )
-}
-
-/** Human-readable summary of an active filter for its chip. */
-function formatFilterSummary(
-  field: FilterField,
-  filter: ActiveFilter
-): { operatorLabel: string; valueText: string } {
-  const operatorLabel = OPERATORS[filter.operator].label
-  const { arity } = OPERATORS[filter.operator]
-
-  if (filter.operator === 'withinHours') {
-    const option = field.options?.find((o) => o.value === filter.values[0])
-    return {
-      operatorLabel,
-      valueText: (option?.label ?? `${filter.values[0]} hours`).replace(
-        /^Last\s+/i,
-        ''
-      ),
-    }
-  }
-
-  let valueText: string
-  if (arity === 2) {
-    valueText = `${filter.values[0]} – ${filter.values[1]}`
-  } else if (arity === 'multi') {
-    valueText =
-      filter.values.length > 2
-        ? `${filter.values.length} selected`
-        : filter.values.join(', ')
-  } else {
-    valueText = filter.values[0] ?? ''
-  }
-
-  if (field.unit && field.type === 'number') {
-    valueText = `${valueText} ${field.unit}`
-  }
-
-  return { operatorLabel, valueText }
-}
-
-interface FilterChipProps {
-  field: FilterField
-  filter: ActiveFilter
-  configName: string
-  onChange: (draft: FilterDraft) => void
-  onRemove: () => void
-}
-
-/** A single active filter, rendered as an editable, removable chip. */
-function FilterChip({
-  field,
-  filter,
-  configName,
-  onChange,
-  onRemove,
-}: FilterChipProps) {
-  const [open, setOpen] = useState(false)
-  const { operatorLabel, valueText } = formatFilterSummary(field, filter)
-  const FieldIcon = field.icon
-
-  return (
-    <div className="inline-flex h-7 max-w-full items-center rounded-md border bg-card pl-2 pr-0.5 text-xs">
-      <Popover open={open} onOpenChange={setOpen}>
-        <PopoverTrigger asChild>
-          <button
-            type="button"
-            className="inline-flex items-center gap-1.5 rounded-sm outline-none focus-visible:ring-1 focus-visible:ring-ring"
-          >
-            {FieldIcon && (
-              <FieldIcon
-                className="size-3.5 text-muted-foreground"
-                aria-hidden
-              />
-            )}
-            <span className="font-medium">{field.label}</span>
-            <span className="text-muted-foreground">{operatorLabel}</span>
-            <span className="max-w-[120px] truncate font-semibold text-primary sm:max-w-[180px]">
-              {valueText}
-            </span>
-          </button>
-        </PopoverTrigger>
-        <PopoverContent align="start" className="w-auto p-3">
-          <FilterEditor
-            field={field}
-            configName={configName}
-            initial={{ operator: filter.operator, values: filter.values }}
-            onSubmit={(draft) => {
-              onChange(draft)
-              setOpen(false)
-            }}
-            onCancel={() => setOpen(false)}
-          />
-        </PopoverContent>
-      </Popover>
-      <button
-        type="button"
-        onClick={onRemove}
-        aria-label={`Remove ${field.label} filter`}
-        className="ml-1 rounded p-1 text-muted-foreground hover:bg-muted hover:text-foreground"
-      >
-        <XIcon className="size-3" />
-      </button>
-    </div>
-  )
-}
-
-interface AddFilterPopoverProps {
-  schema: FilterSchema
-  activeKeys: string[]
-  configName: string
-  onAdd: (key: string, draft: FilterDraft) => void
-}
-
-/** Two-step popover: pick a field, then edit its filter condition. */
-function AddFilterPopover({
-  schema,
-  activeKeys,
-  configName,
-  onAdd,
-}: AddFilterPopoverProps) {
-  const [open, setOpen] = useState(false)
-  const [selectedField, setSelectedField] = useState<FilterField | null>(null)
-
-  const availableFields = schema.fields.filter(
-    (field) => !activeKeys.includes(field.key)
-  )
-
-  const handleOpenChange = (next: boolean) => {
-    setOpen(next)
-    if (!next) setSelectedField(null)
-  }
-
-  return (
-    <Popover open={open} onOpenChange={handleOpenChange}>
-      <PopoverTrigger asChild>
-        <Button
-          variant="outline"
-          size="sm"
-          className="h-7 gap-1 border-dashed text-xs"
-        >
-          <PlusIcon className="size-3.5" />
-          Add filter
-        </Button>
-      </PopoverTrigger>
-      <PopoverContent align="start" className="w-auto p-0">
-        {selectedField ? (
-          <div className="p-3">
-            <FilterEditor
-              field={selectedField}
-              configName={configName}
-              onSubmit={(draft) => {
-                onAdd(selectedField.key, draft)
-                handleOpenChange(false)
-              }}
-              onCancel={() => setSelectedField(null)}
-            />
-          </div>
-        ) : (
-          <Command className="w-56">
-            <CommandInput placeholder="Filter by..." className="h-9" />
-            <CommandList>
-              <CommandEmpty>All filters are active</CommandEmpty>
-              <CommandGroup>
-                {availableFields.map((field) => {
-                  const FieldIcon = field.icon
-                  return (
-                    <CommandItem
-                      key={field.key}
-                      value={field.label}
-                      onSelect={() => setSelectedField(field)}
-                    >
-                      {FieldIcon && (
-                        <FieldIcon className="size-4 text-muted-foreground" />
-                      )}
-                      {field.label}
-                    </CommandItem>
-                  )
-                })}
-              </CommandGroup>
-            </CommandList>
-          </Command>
-        )}
-      </PopoverContent>
-    </Popover>
-  )
-}
-
-interface PresetsMenuProps {
-  presets: FilterPreset[]
-  onApply: (preset: FilterPreset) => void
-}
-
-/** Dropdown of one-click filter bundles. */
-function PresetsMenu({ presets, onApply }: PresetsMenuProps) {
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button variant="outline" size="sm" className="h-7 gap-1 text-xs">
-          <SparklesIcon className="size-3.5" />
-          Presets
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="start" className="w-52">
-        {presets.map((preset) => {
-          const PresetIcon = preset.icon
-          return (
-            <DropdownMenuItem
-              key={preset.name}
-              onClick={() => onApply(preset)}
-              className="gap-2"
-            >
-              {PresetIcon && (
-                <PresetIcon className="size-4 text-muted-foreground" />
-              )}
-              {preset.name}
-            </DropdownMenuItem>
-          )
-        })}
-      </DropdownMenuContent>
-    </DropdownMenu>
   )
 }
